# Trn 638 fetcher one shot