https://jira2.iparadigms.com/browse/TRN-638:

Similar to TRN-606. Add a facility to allow fetcher to run until there are no new domains to process, then exit. This will allow our tests to become deterministic (or close to that)