@pchristopher1275 2015-01-13
In terms of timing: net timings of all fetcher tests (averaged over several runs) looked like 

new code: 14.5s
old code: 36.5s

Of the 25 fetcher tests, all but 5 of them became deterministic. The 5 timed tests had timing intervals that summed to about 6.5s. This means that the 20 untimed functions used to take about 30s, now take about 8s.